New comments

+++ social-media9.Api/Controllers/CommentsController.cs
@@ -3,8 +3,8 @@ using MediatR;
 using social_media9.Api.Models;
 using social_media9.Api.Commands;
 using social_media9.Api.Dtos;
-
-
+using System.Security.Claims;
+using social_media9.Api.Repositories.Interfaces;
 
 namespace social_media9.Api.Controllers
 {
@@ -13,10 +13,14 @@ namespace social_media9.Api.Controllers
     public class CommentsController : ControllerBase
     {
         private readonly IMediator _mediator;
+        private readonly IUserRepository _userRepository;
+        private readonly ICommentRepository _commentRepository;
 
-        public CommentsController(IMediator mediator)
+        public CommentsController(IMediator mediator, IUserRepository userRepository, ICommentRepository commentRepository)
         {
             _mediator = mediator;
+            _userRepository = userRepository;
+            _commentRepository = commentRepository;
         }
 
         [HttpPost]
@@ -37,6 +41,10 @@ namespace social_media9.Api.Controllers
         [HttpDelete("{postId}/{commentId}")]
         public async Task<IActionResult> DeleteComment(Guid postId, Guid commentId)
         {
+            if (!IsUserAuthorized(postId,commentId))
+            {
+                return Unauthorized("You are not authorized to delete this comment.");
+            }
             await _mediator.Send(new DeleteCommentCommand(commentId, postId));
             return NoContent();
         }
@@ -44,6 +52,11 @@ namespace social_media9.Api.Controllers
         [HttpPut("update")]
         public async Task<IActionResult> UpdateComment([FromBody] UpdateCommentDto dto)
         {
+            if (!IsUserAuthorized(dto.PostId,dto.CommentId))
+            {
+                return Unauthorized("You are not authorized to update this comment.");
+            }
+            
             var command = new UpdateCommentCommand
             {
                 CommentId = dto.CommentId,
@@ -54,5 +67,22 @@ namespace social_media9.Api.Controllers
             var result = await _mediator.Send(command);
             return result ? Ok("Updated") : BadRequest("Update failed");
         }
+        public bool IsUserAuthorized(Guid postId,Guid commentId)
+        {
+                        
+            string? googleId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            string userId = _userRepository.GetUserByGoogleIdAsync(googleId).Result?.UserId ?? string.Empty;
+            var comment = _commentRepository.GetCommentByIdAsync(postId,commentId).Result;
+            if (comment == null)
+            {
+                return false; 
+            }
+            else
+            {
+                return comment.UserId == userId ;
+            }
+
+        }
     }
+
 }
